--- Rubberduck.Main/ITestEngineProvider.cs
@@ -1,6 +1,0 @@
-﻿using Rubberduck.UnitTesting;
-
-namespace Rubberduck
-{
-
-}
