+++ Rubberduck.UnitTesting/UnitTesting/ITestEngine.cs
@@ -1,6 +1,5 @@
 ﻿using System;
 using System.Collections.Generic;
-using static System.Net.Mime.MediaTypeNames;
 
 namespace Rubberduck.UnitTesting
 {
